Move truthtables CustomOp into logicnets folder
